Allow REST server to run as user and acquire temporary elevated privileges for pairing.
When running the REST server as root instead of as regular user (so that automatic pairing works) pw-play no longer works because pipewire runs as a user service.
Now also supporting running the REST server as usual but for pairing we run a custom python script (requires elevated priviledges for python3 exe) that then invokes the pairing via D-Bus.