Improve entity optimistic update vs coordinator update by introducing an update counter per coordinator data entry retrieved from the device.
The coordinator now maintains a _data_counters dictionary for each data entry retrieved from the device. When a coordinator update is started a copy of this is made. At the end of the update, which can take a bit, the counter values are compared to the values at the start of the update. If in the meantime an entity value changed, e.g. via user input by flipping a switch, that now calls the new async_update_data_optimistic function on the coordinator, which increments the data entry counter. The coordinator update at the end will observe that the counters don't match and will not update the mismatched data entry from what it received from the device, but instead use the optimistic update value from the entity. It then will trigger an immediate device update right after to get the correct device state.

This should fix issues where when an entity modified device state at just the right time concurrently with the coordinator update, the entity value may flip flop because the coordinator update would push stale data back to the entities.