I've noticed that the device state sometimes after the initial connection doesn't match the state the device reports to the initial coordinator update.
So now we push the coordinator's live data to the device right after the first coordinator update after initial connection to the device.
This should only happen if that initial coordinator update wasn't interrupted via an optimistic update. If it was, we already trigger another immediate coordinator update, and then we do it at that time (again assuming no optimistic interruption).

The coordinator data state pushed on the device when this occurs is:
- volume
- eye_icon
- action
- lights